Refactor all the package
The package until now has been a lot disordered.
Now all have a place to stay and there is order inside the project.
The project is almost ready at first deploy.
