+++ demoMurPy.py
@@ -1,6 +1,10 @@
-from _environment import Environment
-from _commands import VAR, SET, ADD, SUB, MUL, IF, ELSE, ENDIF
 from BFVM import brainfuck
+
+from murpy.commands.memory.stack import VAR, SET
+from murpy.commands.operators.math import ADD, SUB, MUL
+from murpy.commands.controlflow.IF import IF, ELSE, ENDIF
+
+from murpy.compiletools import Environment
 
 
 def main_old():
@@ -26,6 +30,7 @@ def main():
     ENDIF()
     VAR("C", 3)
     # 1 0 3 | reg
+
 
 if __name__ == '__main__':
     env = Environment()
