+++ out.bf
@@ -1,1 +1,1 @@
-+++++>>>[-]<<<[->>>+>+<<<<]>>>>[-<<<<+>>>>][-]+>[-]<<[>>>[-]<<<<<[->>>>>+>+<<<<<<]>>>>>>[-<<<<<<+>>>>>>][-]+>[-]<<[<<<<<[-]++++>>>>>>-<[>>+<<-]]>>[<<+>>-]<[<<<<<<<[-]+>>>>>>>-]<<<-<[>>+<<-]]>>[<<+>>-]<[<<<[-]+>>>-]<<+++
++++++>>>[-]<<<[->>>+>+<<<<]>>>>[-<<<<+>>>>][-]+>[-]<<[>>>[-]<<<<<[->>>>>+>+<<<<<<]>>>>>>[-<<<<<<+>>>>>>][-]+>[-]<<[<<<<<[-]++++>>>>>>-<[>>+<<-]]>>[<<+>>-]<[<<<<<<<[-]+>>>>>>>-]<[-]>[-]>[-]<<<<-<[>>+<<-]]>>[<<+>>-]<[<<<[-]+>>>-]<[-]>[-]>[-]<<<+++
